Create TrainingFormComponentContainer and update Readme
