+++ 02 Sample App/05 EditTraining/src/pages/training/edit/page.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import {Training} from '../../../models/training';
-import {TrainingFormComponent} from './components/trainingForm';
+import {TrainingFormComponentContainer} from './components/trainingFormContainer';
 
 interface Props {
   training: Training;
@@ -12,7 +12,7 @@ export const TrainingEditPage = (props: Props) => {
   return (
     <div>
       <h2 className="jumbotron">Edit Training</h2>
-      <TrainingFormComponent
+      <TrainingFormComponentContainer
         training={props.training}
         onChange={props.onChange}
         save={props.save}
